Create Analyzing the Github History of the Scala Language
Scala is also an open source project. It is a general-purpose programming language that has recently become another prominent language for data scientists.  In this project, We will find out who has had the most influence on its development and who are the experts.

Files can obtained from Kaggle or datacamp.